requirements.txt
django==1.10
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.3
markdown2==2.3.1
Pillow==3.1.0
psycopg2==2.6.2

